added minilib events

+++ inc/manip.h
@@ -6,7 +6,7 @@
 /*   By: sqiu <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/09 11:36:02 by sqiu              #+#    #+#             */
-/*   Updated: 2023/02/15 17:17:09 by sqiu             ###   ########.fr       */
+/*   Updated: 2023/02/16 14:05:26 by sqiu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@
 */
 
 t_point	vec_add(t_point vec1, t_point vec2);
-t_point	vec_scalarmult(t_point vec, int scalar);
+t_point	vec_scalarmult(t_point vec, double scalar);
 
 #endif
